fix: passthrough Runtime.Values in self-migration
Add explicit C# project passthrough semantics, restore the 56-app translate set, close the companion test translation gaps, and revise ADR-0192 for first-class G# Index/Range expressions.


Copilot-Session: 218ebaa1-9cfb-4a06-a479-f36ee70a63ed

+++ test/Core.Tests/CodeAnalysis/Symbols/Adr0186PlatformTypeSymbolTests.cs
@@ -489,9 +489,9 @@ public class Adr0186PlatformTypeSymbolTests
     {
         var cases = new (TypeSymbol Emitted, byte Expected)[]
         {
-            (TypeSymbol.String, 1),
-            (NullableTypeSymbol.Get(TypeSymbol.String), 2),
-            (PlatformTypeSymbol.Get(TypeSymbol.String), 0),
+            (TypeSymbol.String, (byte)1),
+            (NullableTypeSymbol.Get(TypeSymbol.String), (byte)2),
+            (PlatformTypeSymbol.Get(TypeSymbol.String), (byte)0),
         };
 
         foreach (var (emitted, expectedByte) in cases)
